requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
python-dotenv
requests
